requirements.txt
aioredis==1.2.0
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.3
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
daphne==2.2.5
decorator==4.4.0
Django==2.1.7
django-webpush==0.2.7
django-widget-tweaks==1.4.3
hiredis==1.0.0
http-ece==1.1.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
py-vapid==1.5.0
pycparser==2.19
PyHamcrest==1.9.0
pytz==2018.9
pywebpush==1.6.0
requests==2.21.0
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.2
validators==0.12.5
zope.interface==4.6.0
